PROBLEM: sample_agent
struct 'sample_agent' written by Miska turned out to be a very bad idea
solving a non-problem. Out of five members of this struct, only two are
being used now and these two are: a) a const string passed to the struct
one time and then only being read, never written b) a function pointer
that, in the one place where it's being used, can be called directly.

On top of that, being declared as a global variable, makes it even worse
idea than it already was.

SOLUTION: Remove it.

Signed-off-by: Karol Hrdina <KarolHrdina@Eaton.com>